feat: edit in suggestion content
