+++ frontend/package.json
@@ -61,8 +61,10 @@
     "@tiptap/react": "^2.1.11",
     "@tiptap/starter-kit": "^2.1.11",
     "axios": "^1.5.0",
+    "clsx": "^2.1.1",
     "date-fns": "^2.30.0",
     "formik": "^2.4.4",
+    "i": "^0.3.7",
     "mui-tiptap": "^1.8.3",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
