+++ frontend/src/pages/SuggestedContent/editorref.tsx
@@ -1,0 +1,8 @@
+
+import React from 'react'
+
+
+export interface EditorRef ={
+    getHTMLContent: () => string;
+    // Add other methods or properties if needed
+}
